# server/src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Api Notefy",
		"description": "Uma api para gerenciar suas notas",
		"contact": {
			"name": "Notefy",
			"email": "andrewsantos9113@gmail.com"
		},
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:3333",
			"description": "Local server"
		},
		{
			"url": "https://notefy-api-nwxp.onrender.com",
			"description": "Production server"
		}
	],
	"paths": {
...
		"/create-note/{userId}": {
			"post": {
				"summary": "Cria uma nova nota",
				"description": "Essa rota cria uma nova nota, e retorna o objeto nota. Essa rota é protegida, então é necessário passar o token de autenticação",
				"tags": [
					"Note"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Note"
							},
							"examples": {
								"Andrew": {
									"value": {
										"title": "Nota 1",
										"content": "Descrição da nota 1"
									}
								}
							}
						}
					}
				},
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"description": "Id do usuário",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Nota criada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Note"
								}
							}
						}
					},
					"400": {
						"description": "Erro ao criar nota"
					}
				}
			}
		},
		"/create-note-and-category/{userId}": {
			"post": {
				"summary": "Cria uma nova nota e uma nova categoria",
				"description": "Essa rota cria uma nova nota e uma nova categoria, e retorna o objeto nota e categoria. Essa rota é protegida, então é necessário passar o token de autenticação",
				"tags": [
					"Note"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"content": {
										"type": "string"
									},
									"categoryName": {
										"type": "string"
									}
								}
							},
							"examples": {
								"Andrew": {
									"value": {
										"title": "Nota 1",
										"content": "Descrição da nota 1",
										"categoryName": "Categoria 1"
									}
								}
							}
						}
					}
				},
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"description": "Id do usuário",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Nota e categoria criadas com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Note"
								}
							}
						}
					},
					"400": {
						"description": "Erro ao criar nota e categoria"
					}
				}
			}
		},
...
		"/create-category/{userId}": {
			"post": {
				"description": "Cria uma nova categoria",
				"tags": [
					"Category"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									}
								}
							},
							"examples": {
								"Andrew": {
									"value": {
										"name": "Categoria 1"
									}
								}
							}
						}
					}
				},
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"description": "Id do usuário",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Categoria criada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					},
					"400": {
						"description": "Erro ao criar categoria"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"Note": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"category": {
						"type": "object",
						"properties": {
							"id": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						}
					}
				}
			},
			"Category": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT token"
			}
		}
	}
}